- Merge HomeScreen and MainPage (delete) - Add Save PageSize function
