+++ FoodRecipeApp/GUI/HomeScreen.xaml.cs
@@ -14,8 +14,10 @@ using System.Windows.Navigation;
 using System.Windows.Shapes;
 using MahApps.Metro.Controls;
 using FoodRecipeApp.ViewModels;
-using MenuItem = FoodRecipeApp.ViewModels.MenuItem;
 using System.Configuration;
+using Telerik.Windows.Controls;
+using FoodRecipeApp.DTO;
+using FoodRecipeApp.DAO;
 
 namespace FoodRecipeApp.GUI
 {
@@ -24,99 +26,148 @@ namespace FoodRecipeApp.GUI
 	/// </summary>
 	public partial class HomeScreen : MetroWindow
 	{
-        public static HomeScreen homeScreen;
-
-		private readonly Navigation.NavigationServiceEx _navigationServiceEx;
+		public static HomeScreen AppMainpage;
 		public HomeScreen()
 		{
-            homeScreen = this;
 			InitializeComponent();
-            this._navigationServiceEx = new Navigation.NavigationServiceEx();
-            this._navigationServiceEx.Navigated += this.NavigationServiceEx_OnNavigated;
-            this.HamburgerMenuControl.Content = this._navigationServiceEx.Frame;
+			AppMainpage = this;
+			var viewModel = new RecipeViewModel();
+			this.DataContext = viewModel;
+		}
 
-            // Navigate to the home page.
-            this.Loaded += (sender, args) => this._navigationServiceEx.Navigate(new Uri("GUI/MainPage.xaml", UriKind.RelativeOrAbsolute));
-        }
-
-        private void HamburgerMenuControl_OnItemInvoked(object sender, HamburgerMenuItemInvokedEventArgs e)
-        {
-			if (e.InvokedItem is MenuItem menuItem && menuItem.IsNavigation)
-            {
-                UpdateKey();
-                this._navigationServiceEx.Navigate(menuItem.NavigationDestination);
-            }
-        }
-
-        private void NavigationServiceEx_OnNavigated(object sender, NavigationEventArgs e)
-        {
-            // select the menu item
-            this.HamburgerMenuControl.SelectedItem = this.HamburgerMenuControl
-                                                         .Items
-                                                         .OfType<MenuItem>()
-                                                         .FirstOrDefault(x => x.NavigationDestination == e.Uri);
-            this.HamburgerMenuControl.SelectedOptionsItem = this.HamburgerMenuControl
-                                                                .OptionsItems
-                                                                .OfType<MenuItem>()
-                                                                .FirstOrDefault(x => x.NavigationDestination == e.Uri);
-
-            // or when using the NavigationType on menu item
-            // this.HamburgerMenuControl.SelectedItem = this.HamburgerMenuControl
-            //                                              .Items
-            //                                              .OfType<MenuItem>()
-            //                                              .FirstOrDefault(x => x.NavigationType == e.Content?.GetType());
-            // this.HamburgerMenuControl.SelectedOptionsItem = this.HamburgerMenuControl
-            //                                                     .OptionsItems
-            //                                                     .OfType<MenuItem>()
-            //                                                     .FirstOrDefault(x => x.NavigationType == e.Content?.GetType());
-
-            // update back button
-            this.GoBackButton.Visibility = this._navigationServiceEx.CanGoBack ? Visibility.Visible : Visibility.Collapsed;
-        }
-
-        private void GoBackButton_OnClick(object sender, RoutedEventArgs e)
+		static HomeScreen()
 		{
-            UpdateKey();
-            this._navigationServiceEx.GoBack();
-        }
+			var addFavouriteItemBinding = new CommandBinding(TileViewCommandsExtension.AddNewFavouriteRecipe, OnAddFavouriteItemCommandExecute, OnCanAddFavouriteItemCommandExecute);
+			CommandManager.RegisterClassCommandBinding(typeof(RadTileViewItem), addFavouriteItemBinding);
+		}
 
-        //private MetroWindow _accentThemeTestWindow;
-
-        //private void ChangeAppStyleButtonClick(object sender, RoutedEventArgs e)
-        //{
-	       // if (_accentThemeTestWindow != null)
-	       // {
-		      //  _accentThemeTestWindow.Activate();
-		      //  return;
-	       // }
-
-	       // _accentThemeTestWindow = new AccentStyleWindow
-	       // {
-		      //  Owner = this
-	       // };
-	       // _accentThemeTestWindow.Closed += (o, args) => _accentThemeTestWindow = null;
-	       // _accentThemeTestWindow.Left = this.Left + this.ActualWidth / 2.0;
-	       // _accentThemeTestWindow.Top = this.Top + this.ActualHeight / 2.0;
-	       // _accentThemeTestWindow.Show();
-        //}
-
-        public static void UpdateKey()
+		private static void OnCanAddFavouriteItemCommandExecute(object sender, CanExecuteRoutedEventArgs e)
 		{
-            var config = ConfigurationManager.OpenExeConfiguration(ConfigurationUserLevel.None);
-            var viewmodel = new RecipeViewModel();
+			e.CanExecute = true;
+		}
 
-            string AllRecipesPageSize = viewmodel.AllRecipesCustomPageSize.ToString();
-            string FavouriteRecipesPageSize = viewmodel.FavouriteRecipesCustomPageSize.ToString();
+		private static void OnAddFavouriteItemCommandExecute(object sender, ExecutedRoutedEventArgs e)
+		{
+			var tileViewItem = sender as RadTileViewItem;
+			var tileView = tileViewItem.ParentTileView as RadTileView;
 
-            if (!config.AppSettings.Settings["AllRecipesPageSize"].Value.Equals(AllRecipesPageSize))
-            {
-                config.AppSettings.Settings["AllRecipesPageSize"].Value = AllRecipesPageSize;
-            }
-            if (!config.AppSettings.Settings["FavouriteRecipesPageSize"].Value.Equals(FavouriteRecipesPageSize))
-            {
-                config.AppSettings.Settings["FavouriteRecipesPageSize"].Value = FavouriteRecipesPageSize;
-            }
-            config.Save(ConfigurationSaveMode.Minimal);
-        }
+			if (tileViewItem == null || tileView == null || tileView.ItemsSource == null) return;
+
+			var dataItem = tileView.ItemContainerGenerator.ItemFromContainer(tileViewItem) as Dish;
+
+			if (dataItem == null) return;
+
+			if (DishDAO.Instance.updateFavouriteRecipe(dataItem.DishCode.ToString()) == 1)
+			{
+				//Success
+
+				var viewModel = (RecipeViewModel)AppMainpage.DataContext;
+				if (dataItem.IsLove)
+				{
+					var item = viewModel.Recipes.FirstOrDefault(i => i.DishCode == dataItem.DishCode);
+					viewModel.AddNewItemToFavouriteRecipesList(dataItem);
+				}
+				else
+				{
+					var item = viewModel.FavouriteRecipes.FirstOrDefault(i => i.DishCode == dataItem.DishCode);
+					viewModel.RemoveItemFromFavouriteRecipesList(item);
+					foreach (var tom in viewModel.Recipes.Where(w => w.DishCode == dataItem.DishCode))
+					{
+						tom.IsLove = dataItem.IsLove;
+					}
+				}
+				/*if (!viewModel.UpdateFavouriteRecipesCustomPageSize())
+				{
+					MessageBox.Show("Error");
+				}*/
+			}
+		}
+
+		private void radTileView_TileStateChanged(object sender, Telerik.Windows.RadRoutedEventArgs e)
+		{
+			RadTileViewItem item = e.OriginalSource as RadTileViewItem;
+
+			if (item == null) return;
+
+			RadFluidContentControl fluid = item.ChildrenOfType<RadFluidContentControl>().FirstOrDefault();
+
+			if (fluid == null) return;
+
+			switch (item.TileState)
+			{
+				case TileViewItemState.Maximized:
+					fluid.State = FluidContentControlState.Large;
+					break;
+				case TileViewItemState.Minimized:
+					fluid.State = FluidContentControlState.Normal;
+					break;
+				case TileViewItemState.Restored:
+					fluid.State = FluidContentControlState.Normal;
+					break;
+				default:
+					break;
+			}
+		}
+
+		private void DetailSteps_Click(object sender, RoutedEventArgs e)
+		{
+			Button temp = (Button)sender;
+			int DishCode = (int)(temp.Tag);
+			List<Step> steps = Step.getAllStepsInDish(DishCode);
+			var addedStepsScreen = new AddedStepWindow(steps);
+			addedStepsScreen.Dying += OpenThis;
+			addedStepsScreen.Show();
+			//HomeScreen.homeScreen.Hide();
+		}
+
+		private void OpenThis()
+		{
+			///HomeScreen.homeScreen.Show();
+		}
+
+		private void Pager_PageIndexChanging(object sender, PageIndexChangingEventArgs e)
+		{
+			//Do nothing
+		}
+
+		private void VideoDishButton_Click(object sender, RoutedEventArgs e)
+		{
+			Button temp = (Button)sender;
+			string Video = temp.Tag.ToString();
+			youtubeWindow youtubeWindow = new youtubeWindow(Video);
+			youtubeWindow.Show();
+		}
+
+		private void MainTabControl_SelectionChanged(object sender, SelectionChangedEventArgs e)
+		{
+			if (e.Source is MetroAnimatedSingleRowTabControl)
+			{
+				//do work when tab is changed
+				//UpdateKey();	
+			}
+		}
+
+		private void RadFluidContentControl_MouseDoubleClick(object sender, MouseButtonEventArgs e)
+		{
+
+		}
+
+		private void MetroWindow_Closed(object sender, EventArgs e)
+		{
+			string AllRecipesPageSize = AllRecipesNumericUpDown.Value.ToString();
+			string FavouriteRecipesPageSize = FavoriteNumericUpDown.Value.ToString();
+
+			var config = ConfigurationManager.OpenExeConfiguration(ConfigurationUserLevel.None);
+
+			if (!config.AppSettings.Settings["AllRecipesPageSize"].Value.Equals(AllRecipesPageSize))
+			{
+				config.AppSettings.Settings["AllRecipesPageSize"].Value = AllRecipesPageSize;
+			}
+			if (!config.AppSettings.Settings["FavouriteRecipesPageSize"].Value.Equals(FavouriteRecipesPageSize))
+			{
+				config.AppSettings.Settings["FavouriteRecipesPageSize"].Value = FavouriteRecipesPageSize;
+			}
+			config.Save(ConfigurationSaveMode.Minimal);
+		}
 	}
 }
